requirements.txt
selenium
pytest
webdriver-manager
pytest-html
python-dotenv